config: Intentionally throw TypeError for unsupported sources
For config sources that are not supported i.e. not either of the
following (Config, ServiceOptions or array), we should fail hard
as we don't want to allow anything else apart from those.

`array_key_exists()` would throw a TypeError if a source is `null`
for example which would be the appropriate behavior here since we
don't support that.

This was introduced in I06d16b473b256d3f9 (3e495f35c4d0ac6) and
a simpler fix to the PSR-12 warning is not to do the `if(..){}` inside
an `else`, but use `elseif(..){}` instead.

Change-Id: Ib928871b6b4cefb6872484437ecb4c04a9327bad