SUL3: Enable multiple auth providers to refresh "remember me" status
This allows a primary auth providers to update the "remember me" status
of another primary provider in the same login request flow. If authentication
that happened elsewhere and a "remember me" / "keep me logged in" (extended
login period is selected), the value of this action applied by the remote
primary provider should be applied to the local wiki when the user is
redirected back to continue authentication there.

This is useful because:

In the case of Wikimedia's central domain wiki in SUL3 mode, we want to
apply the remember me (keep me logged in) flag value from the central domain
to the local domain.

Bug: T369668
Change-Id: I6e2e2d892d2b777cb4757c7c0d222afad9da506c