CommentStore: `hash()` has only one caller, make private
This method seems to have only a single caller which happens to be
in the CommentStore.php class itself. Maybe we should just make this
a private utility.

Test plan
=========

If this patch fails, then I'm probably wrong and something else is
using it otherwise, then we can make it private.

Change-Id: Ib774af48ddf7e0a42a1c9a08cdecd5317e238cb4