Restore password reset invalid email message
This mostly reverts commit 28ad29ea0ff9480c97d4d5746f5b2526dfb1a3f5.

Checking the validity of the provided email doesn't reveal any
information about any users, and it could help catch genuine mistakes,
e.g. if someone inputs the username in the email field.

Previously if an invalid email was provided, we showed a success
message that is an obvious lie (saying we sent an email to something
that isn't an email address).

Change-Id: Idec437b29b22e5b1e5aaff3846632fbc413a3dcb