Remove ParserOutput::getText() calls from core (getRawText)
This is the first patch of a series of patches to remove
ParserOutput::getText() calls from core. This series of patches should
be functionally equivalent to I2b4bcddb234f10fd8592570cb0496adf3271328e.

This first patch replaces the calls to getText to calls to
ParserOutput::getRawText when the pipeline has no reason to be executed.

Bug: T293512
Change-Id: I0ad53cd074ca9cf13e96e6b08179e13aeea180f4