deferred: actually flush replica view snapshots in onRunUpdateStart()
The comment mentioned this, but the code did not actually do it.

Typically, flushing these does not matter much due to replication
lag. However, if a deferred update starts with a replication wait
or the prior one ended with a replication wait, then flushing the
read transactions can make a reliable difference.

Also, if many deferred updates are running in a process, flushing the
snapshots means that mysql can purge old versions of rows, since
otherwise they could potentially be seen by a future query (e.g. MVCC).

Bug: T315664
Change-Id: Ia3ac4193e56ef5b49727457b0cd50dfbc987f881