Made the JsonContent Responsive
Wrapped the HTML of JsonContent with a div and added a custom class to it. The class .noresize in mediawiki.content.json.less sets the overflow-x to auto and white-space to nowrap. This will make the JsonContent responsive with a horizontal scroll bar. Updated tests for the new behaviour of the JsonContent.

Bug: T366300
Change-Id: Ibf1494fa0940c37cf36ebfa18442d86c318dd1d1