parser: Optimize heading regex for backtracking performance
This regex can apparently run into the backtracking limit. The ticket
T364854 mentions a vandalized old revision with thousands of headings.
We can completely avoid half of the backtracking without changing any
behavior.

Why?

The previous .*?> was an ungreedy match (the extra ? makes it
ungreedy) for any number of characters. But the > at the end makes it
stop at the very first > character it can find. Because of this the
.*? part can never contain a > character.

The new regex does the same as before, but explicitely says that, and
can never run into any backtracking because of this.

I'm also replacing the [\s\S] trickery with the more straightforward
/s modifier that has the same effect. This part will still backtrack.

Capitalization doesn't matter because of the /i modifier. I change it
to make it more consistent with all other code where we write <h1> in
lowercase.

The \b is added to make it ignore irrelevant sequences like <h11>.

Bug: T364854
Change-Id: Id5c02ebe4fc05b27d6f270805296c619b79ada49