Make DeletePage use commitAndWaitForReplication()
This is the preferred way to wait for replicas to catch up, where
as using direct IDatabase::commit() calls is discouraged. Note that
jobs already run with "outer transaction scope". Remove the warning
code since commitAndWaitForReplication() already logs and bails out
if there is a transaction that cannot be flushed by the caller.

This also fixes a bug where a "continue" caused by the check to
explicitTrxActive() would result in mismatched endAtomic() calls.

Change-Id: Ic87ddf16cce810ae60b1f54426671a7a1509dcef