Use Codex provided icon and color in HTMLCheckMatrix
The only use of images/questions.svg apparently was in
HTMLCheckMatrix and this removes the use making deletion
of the asset also possible.

As this uses a Codex provided color, it makes the help icon
to now look correctly in the dark mode also.

This is done in a way that HTMLCheckMatrix's clients can
provide their own style to override the icon like before.

The main motivation behind this is to remove inaccurate
:lang() selectors this had for the help icon's localization which
was supposed to support different languages that use an RTL script
but the Codex one also already provides the same and more
complete logic automatically using the Less mixin it provides.

Change-Id: If5d11fd9c6fc9d528f6792d2d35fa097c89fba5a