RL: Consider RTL styles are also generated
ResourceLoader currently declares styles are generated only if
Less is involved but RTL CSS styles are also generated in the
same way so let's treat them the same.

This makes RL's debug mode matching better with non-debug mode
so if a user uses ?debug=1 in an RTL wiki, they get CSS Janus
transformed styles just as if Less was involved.

I was seeing the issue for years but had to get to the bottom
of it in I6c9544d7eee3b5885f4fd30265e906fe22a8c223 as this
bit was part of the confusion I had to debug the issue there.

Change-Id: Ib10bec85da969881be590e3d95d5f9474f2eb6e9