Stats: add special bucket attribute to MetricTrait
A bucket is a special label used by Prometheus histogram query
functions.  It requires key/values that would be mangled by the
normalizer if treated as a normal label.

StatsD format is incompatible with Prometheus-style histogram
definitions, so the formatter will drop the metric entirely to try and
prevent confusing metrics from being generated.

Bug: T348796
Change-Id: Iba4a080e7de8ae242964b33f8b3b07f13d70d411