Make "View contributions" menu item the default, instead of "Contribute"
Where the "Contribute" menu is available (for example, in Albanian
Wikipedia), it provides users with two tabs: "Contribute" (with
access to tools to start new contributions) and "View Contributions"
(the list of past contributions by the user). Currently, in these
wikis the "Contribute" tab is always the default. Thus, users who
regularly access "View contributions" as part of their workflow, have
to always switch to the "View Contributions" tab.

To avoid this issue, we want to adjust the navigation menu option to
point to the "View contributions" tab, when the user has at least
one edit.

This patch implements the above logic.

Bug: T369041
Change-Id: Iddd8fe8c9f36781adf6674a3587b99f7ce3c9e45