Allow change tags to have help links in separate messages
Many changes tags added by extensions include links in the 'tag-<tag>'
label messages, which point to help pages for the tags and the software
features that add them. This makes translating the labels unpleasant.

Introduce localisation messages like 'tag-<tag>-helppage' (following
the pattern of 'tag-<tag>-description'), which should contain a URL
or a page title of a help page for the given tag. When such link
message is defined, the tag description in changes lists, history
pages etc. is wrapped in a link that points to the given URL or page.

Unlike the label messages, the link messages are always used in wiki
content language, for easy on-wiki customization. For compatibility
with old translations and existing on-wiki overrides, if 'tag-<tag>'
already contains any links, 'tag-<tag>-helppage' is ignored.

Add buttons that allow editing the new links on Special:Tags. Correct
existing buttons that allow editing the labels to use user language.

Add help links using the new messages to the changes list filtering
interface (which can't display links in tag descriptions themselves).

Introduce a link message for 'mw-contentmodelchange' tag as an example.

Bug: T372175
Change-Id: I1e2242a8591597a110a5a430f4c251f0f601cfbe