Stats: implement HistogramMetric
Implements HistogramMetric as a collection of counters with the
correct suffixes and labels necessary to enable the use of histogram
functions in Prometheus.

TimingMetric is limited to the default set of buckets configured by
statsd-exporter and is oriented toward measuring units of time in
seconds.  HistogramMetric allows the user to define the buckets with
the metric definition and is not limited to measuring units of time.

The statsd formatter is incompatible with HistogramMetrics - these
metrics will be dropped when setting $wgStatsFormat = 'statsd'.

TimingMetrics migrating to HistogramMetric require a statsd-exporter
restart, otherwise statd-exporter will report a conflict and the
observation will not be reported and the conflict counter incremented.
This can also be worked around if the metric is renamed.

Bug: T348796
Change-Id: I87a2574394c6d86283cc371305626ef0d8f697e1