installer: Support MW_SKIP_EXTERNAL_DEPENDENCIES in update.php
This environment variable is already checked in TestSetup.php for
PHPUnit runs by Quibble, when the job is running for a patch to
the mediawiki/vendor repository.

Right now, Quibble sets this var, but then has to separately pass
`--skip-external-dependencies` to update.php as well.

Remove the need for that by supporting the same environment variable
so that Quibble can set it once for that run, and naturally applies
by any commands where it makes sense.

Bug: T370380
Change-Id: Iad7c8d4be9e81af0a0de11019a98804b232efa8b
(cherry picked from commit 4c41313130468123884da1381013f07a6f0c8de0)