Clarify precedence between conflicting localized image parameter names
For an image parameter like 'maldekstra' in Esperanto, the reason it is
parsed as `img_left` instead of `img_width` (`$1ra`) is due to some
subtle ordering of parameter names in Parser::getImageParams().  Document
this behavior at the same time we move `img_manualthumb` and
`img_upright` a little bit backward in the list to help avoid similar
conflicts in the future.

Bug: T372935
Change-Id: I5dec0e32d98bb931db267650d7183d8115db8d0e