rdbms: remove CONN_INTENT_WRITABLE and deprecate other CONN_* constants
This was only used to make primary connection count logging show for LBs
with a single DB server. It did not fully work since a prior connection
for a DB_REPLICA handle would mask any warnings for a later DB_PRIMARY
handle due to connection reuse.

Instead of this, just make the recordConnection() parameter operate on
whether there are replicas or not. Various scripts, vagrant roles, and
docker compose recipies make it easy to setup replication, if someone
really wants to check these TransactionProfiler logs.

Change-Id: Ied425a6132489c2f1feeba779e86dd5e7a262ba3