docs: Set SORT_GROUP_NAMES=YES in Doxyfile
The current order is decided based on whatever order Doxygen
encountered the `@defgroup myname Title Here` commands when
recursively traversing the source code directories (INPUT).

By default, it is not based on the identifier (group name) or display
label (group title), which means we end up with the following
on doc.wikimedia.org:

* API
* Recent changes (!)
* Change tagging
* Debug
* ExternalStorage
* Profiling
* ExtensionRegistry (!)
* ResourceLoader
* Upload
* Maintenance (!)

Because internally these are traversed as:

* includes/api
* includes/changes
* includes/registration
* maintenance/

But that's by no means known, relevant, or intuitive to a reader
of the docs; even for contributors that know where the files are,
they probably don't realize that maintenance/ comes after all
includes.

Bug: T364652
Change-Id: I3f03ae5bcc902aa2df020b7d5aa404c42829320c