More robust assert.step() call in mediawiki.base/track test
This test was added not long ago, see Ie5b8ae4. We see it failing in
unrelated patches, see I7dc9e26. Unfortunately the error message is
not helpful because assert.step() apparently ends being called with
incompatible arguments. This will probably not fix the issue, but
make the output more helpful.

Needed-By: I7dc9e26edc954d195dff8aeefa7d923ea7534fea
Change-Id: I75d74e74eca0bdfb592e2d3ababcaea062d09190