ResourceLoader: Fix `require('../../foo.js')` from module base directory
Follows-up 657872dee50e9, which fixed `../` access to preserve 0 or 1
above the module base directory. It didn't do that intentionally, but
it was a side-effect of the uncached .length check.

Once the array is empty, we reach the base directory, and interpret
`..` as something to push instead of pop. That's all great.

But, then the next `..` would pop again since it is now non-empty.

Fix this by remembering whether we reached the base, and only pushing
from that moment on. Don't switch from pushing back to popping.

Bug: T373065
Change-Id: I777f2c12e845a738edeac00e19818d4c939a1ef1