Fix vertical align of collapsibleFooter toggle across skins
Different skins have different values for <p>'s spacing making toggle
arrow hard to vertically align correctly with heading paragraph, this
resets it to the symmetric spacing of Vector-2010.

Admittedly this isn't that clean solution but the other solution
would be more intervention on the PHP and also LivePreview side,
turning parseAsBlock() to parse() and taking care of Wikibase
which can't be done in rush so let's have a temporarily fix
at least for now.

Bug: T373658
Change-Id: Ie0d8014d7716b11f04c7b0de782b8ce1d15ff0f7