Introduce DomainEventDispatcher
This introduces infrastructure for emitting domain events.
Domain events work similar to hooks, but use event object instead of
a list of parameters, and they are not dispatched immediately but
in a deferred update.

To demonstrate the concept, this patch introduces PageChangeEvent along
with the relevant wiring. This is used to simplify the logic in
PageUpdater.

This change deprecates PageUpdater::setUsePageCreationLog().

Bug: T377229
Change-Id: I48977ac0547bd7bd694397e0561dbb1ad1c9a42e