CleanupTitles: Check if title exists from primary database
It's possible for multiple invalid titles to normalize to the same
title. The script tries to catch this, but it uses a replica database
for it, which might not have picked up an earlier change in the same
maintenance script. This caused the cleanupTitles run on ptwiki to crash
(and somehow escaped notice until now)

Bug: T195546
Change-Id: I2a4a34ce99067986d7c44605c796323c4d77dc4e