mediawiki.special.block [vue]: add new ExpiryField to enter block expiry
This uses a new three-part form to allow the features of manually typing
an expiry (i.e. "31 hours") but allowing to do so for non-English users,
who previously didn't have this capability unless it was set in
MediaWiki:Ipboptions. The ability to use MediaWiki:Ipb-default-expiry to
set a default expiry is still supported, assuming it is either a valid
option in MediaWiki:Ipboptions or in the new "Custom duration" field
(i.e. "31 hours"). See https://w.wiki/B65S for more, though this
functionality doesn't appear to be used anywhere on the WMF cluster.

"Expiration" is renamed in the code to "Expiry" to match the naming of
other MW components (ExpiryDef), but the user-facing message is still
"Expiration".

No validation is done yet, but will be added in a future patch.
Similarly, pre-supplying values via URL params is saved for T368583.

Bug: T359684
Change-Id: Iffb6fa3f539d29ed25802a19f4c13c627208259c