auth: Pass canAlwaysAutocreate from session to pre-auth providers
Session providers can provide a `canAlwaysAutocreate` flag which
indicates account creation is exempt from autocreate permission
checks. This is used, for example, for providers that provide
users for supporting applications in a wiki farm.

Pass that information along to pre-authentication providers so
they can respect the intent of the flag. We cannot simply skip
pre-authentication, it's used for non-permission-related things
like preventing CentralAuth username conflicts.

Bug: T373778
Change-Id: Ie3aeaf48e615e2fb85b2069203ab91ca0127ae05