ParserOutput::collectMetadata(): fix handling of links
For language links, when there are conflicts between namespaces and
interwiki prefixes, it is important to use TitleValue for language links
rather than to try to reparse the Title.  Language links also preserve
fragments, unlike other link types in ParserOutput; added tests to
document this.

Added handling for interwiki links and template links.

Bug: T363538
Change-Id: I6e8ff8ed7f8819000cc3f80e49c0739b568217a4